% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/externals.R
\name{clusterNet}
\alias{clusterNet}
\title{clusterNet performs consensus clustering using an adjacency matrix for the network}
\usage{
clusterNet(data, type = c("adjacency_matrix", "edge_list"), tau = 0.5, ...)
}
\arguments{
\item{data}{A network file, whether it be a correlation matrix or edge list}

\item{type}{a string identifying the type of input for the data parameter. \strong{adjacency_matrix} is used for a
correlation matrix and \strong{edge_list} is used for inputting an edge list. \emph{(see details for additional arguments
required for each type)}}

\item{tau}{The consensus probabilty threshold for agreement among clustering algorithms}
}
\value{
A list containing:

\if{html}{\out{<div class="sourceCode">}}\preformatted{    1. A data frame, named "subnetworks", containing sub-network determinations for the nodes within the input network.

    2. A data frame, named "summary", with a summary of the number of edges and nodes in each subnetwork.
}\if{html}{\out{</div>}}
}
\description{
This function will take as input an adjacency matrix from the determined network and perform
consensus clustering using the following methods from the igraph package: cluster_edge_betweenness,
cluster_fast_greedy, cluster_infomap, cluster_label_prop, cluster_leading_eigen, cluster_louvain,
cluster_walktrap. The output results in sub-network classification for the nodes within the network.
}
\details{
\subsection{Additional Parameters}{

\strong{weighted} \emph{REQUIRED only for type = "adjacency_matrix"} TRUE/FALSE indicating whether the adjacency graph is weighted.

\strong{metaboliteA} \emph{REQUIRED only for type = "edge_list"} A string indicating the column name of the first column in the edge pair.

\strong{metaboliteB} \emph{REQUIRED only for type = "edge_list"} A string indicating the column name of the second column in the edge pair.
}
}
