% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{clusterNetResults}
\alias{clusterNetResults}
\title{clusterNetResults}
\usage{
clusterNetResults(x)
}
\arguments{
\item{x}{clusterNetResults object}
}
\description{
clusterNetResults
}
\details{
clusterNetResults creates a generic for clusterNetResults
}
