% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{correlation_data}
\alias{correlation_data}
\title{The example data from the Correlation Calculator tool}
\format{
\subsection{\code{correlation_data}}{

A data frame with 741 rows and 5 columns:
\describe{
\item{metab1}{metabolite name of first of edge pair}
\item{metab2}{metabolite name of second of edge pair}
\item{pcor}{The partial correlation value of the edge}
\item{pval}{Pvalue associated witht the edge}
\item{adj.pval}{The adjusted pvalue (using Benjamini-Hochberg) associated with the edge}
...
}
}
}
\source{
\url{http://metscape.ncibi.org/files.html}
}
\usage{
correlation_data
}
\description{
An edge list of data from ...
}
\keyword{datasets}
