% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{gatherConsensusMatrix}
\alias{gatherConsensusMatrix}
\title{gatherConsensusMatrix will put together a model.matrix type summary of subnetwork results}
\usage{
gatherConsensusMatrix(consensus_membership)
}
\arguments{
\item{consensus_membership}{A numeric vector of subnetwork classification for each metabolite}
}
\value{
a matrix with subnetworks as rows and metabolites as columns. The values are 0 if the metabolite is NOT
in said subnetwork and a 1 if it is.
}
\description{
The function takes as input the metabolite info and subnetwork classfication to create an expanded matrix
}
\keyword{internal}
