% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{run_consensus_cluster}
\alias{run_consensus_cluster}
\title{Performs consensus clustering}
\usage{
run_consensus_cluster(
  adjacency_graph,
  tau,
  num_iterations = 10,
  maxIter = 5,
  main.seed
)
}
\arguments{
\item{adjacency_graph}{graph An adjacency matrix of the determined network.}

\item{tau}{The consensus probabilty threshold for agreement among clustering runs}

\item{num_iterations}{The total number of iterations to perform.}

\item{maxIter}{Maximum number of iterations to perform.}

\item{main.seed}{An integer to set seed for clustering. This parameter ensures reproduciblity and can be
kept as default.}
}
\value{
Sub-network determinations for the nodes within the input network
}
\description{
This function will take as input an adjacency matrix graph from the determined networks and perform
consensus clustering using the following methods from the igraph package: cluster_edge_betweenness,
cluster_fast_greedy, cluster_infomap, cluster_label_prop, cluster_leading_eigen, cluster_louvain,
cluster_walktrap. The output results in sub-network classification for the nodes within the network.
}
\keyword{internal}
